[fix] Serviceのinterfaceをファイルにまとめた
